refactor(client): Separate logic for add dataset to an existing project from Add dataset container
